Upgrade to LLVM 13:
* Fix GCC 11 compilation issues

* Fix required alignment issues in LLVM 11:

This seems to stem from this conversation:

https://reviews.llvm.org/D77454

This new alignment requirement is new in LLVM 11, since LLVM 10 did not
have this issue.

* Fix llvm-config/llc version resolution

+++ skull/common/hashtable.c
@@ -26,7 +26,7 @@ static __attribute__((pure)) unsigned ht_hash_key(const char *const key) {
 	unsigned long long hash = 0;
 
 	for RANGE(i, len) { // NOLINT
-		hash += (unsigned)key[i] * i;
+		hash += (unsigned long long)key[i] * i;
 	}
 
 	return hash % MAX_SLOTS;
